One best-effort reconcile policy for automatic setup triggers
sail up and the boot sweep each carried an identical try-ensureInstalled-
warn wrapper. The single-container policy now lives once, in
ContainerSetupSweep.reconcile; the sweep loops it and sail up calls it.
Callers needing the outcome or their own error handling (dispatch,
provisioner, renamer, reconfigure) keep calling ensureInstalled directly.
